Practica2/app/requirements.txt
pytest
pytest-cov
beautifulsoup4
io
shutil
numpy
pandas
requests
matplotlib
HTMLResponse
Jinja2
requests
math
statsmodels
pydantic
aiofiles
python-multipart
Markdown
Pillow
uvicorn